Code/test arrangement
