+++ custom_components/winix/WinixDeviceWrapper.py
@@ -1,12 +1,12 @@
 """The Winix C545 Air Purifier component."""
 
-
 from typing import Dict
 
 import aiohttp
 
 from winix import WinixDeviceStub
 
+from .WinixDriver import WinixDriver
 from .const import (
     AIRFLOW_LOW,
     AIRFLOW_SLEEP,
@@ -36,7 +36,7 @@ class WinixDeviceWrapper:
         logger,
     ):
         """Initialize the wrapper."""
-        self._driver = WinixDevice(device_stub.id, client)
+        self._driver = WinixDriver(device_stub.id, client)
 
         # Start as empty object in case fan was operated before it got updated
         self._state = {}
@@ -233,139 +233,3 @@ class WinixDeviceWrapper:
         elif preset_mode == PRESET_MODE_MANUAL_PLASMA:
             await self.async_manual()
             await self.async_plasmawave_on()
-
-
-# Modified from https://github.com/hfern/winix to support async operations
-class WinixDevice:
-    """WinixDevice driver."""
-
-    CTRL_URL = "https://us.api.winix-iot.com/common/control/devices/{deviceid}/A211/{attribute}:{value}"
-    STATE_URL = "https://us.api.winix-iot.com/common/event/sttus/devices/{deviceid}"
-
-    category_keys = {
-        "power": "A02",
-        "mode": "A03",
-        "airflow": "A04",
-        "aqi": "A05",
-        "plasma": "A07",
-        "filter_hour": "A21",
-        "air_quality": "S07",
-        "air_qvalue": "S08",
-        "ambient_light": "S14",
-    }
-
-    state_keys = {
-        "power": {"off": "0", "on": "1"},
-        "mode": {"auto": "01", "manual": "02"},
-        "airflow": {
-            "low": "01",
-            "medium": "02",
-            "high": "03",
-            "turbo": "05",
-            "sleep": "06",
-        },
-        "plasma": {"off": "0", "on": "1"},
-        "air_quality": {"good": "01", "fair": "02", "poor": "03"},
-    }
-
-    def __init__(self, device_id: str, client: aiohttp.ClientSession):
-        """Create an instance of WinixDevice."""
-        self.device_id = device_id
-        self._client = client
-
-    async def turn_off(self):
-        """Turn the device off."""
-        await self._rpc_attr(
-            self.category_keys["power"], self.state_keys["power"]["off"]
-        )
-
-    async def turn_on(self):
-        """Turn the device on."""
-        await self._rpc_attr(
-            self.category_keys["power"], self.state_keys["power"]["on"]
-        )
-
-    async def auto(self):
-        """Set device in auto mode."""
-        await self._rpc_attr(
-            self.category_keys["mode"], self.state_keys["mode"]["auto"]
-        )
-
-    async def manual(self):
-        """Set device in manual mode."""
-        await self._rpc_attr(
-            self.category_keys["mode"], self.state_keys["mode"]["manual"]
-        )
-
-    async def plasmawave_off(self):
-        """Turn plasmawave off."""
-        await self._rpc_attr(
-            self.category_keys["plasma"], self.state_keys["plasma"]["off"]
-        )
-
-    async def plasmawave_on(self):
-        """Turn plasmawave on."""
-        await self._rpc_attr(
-            self.category_keys["plasma"], self.state_keys["plasma"]["on"]
-        )
-
-    async def low(self):
-        """Set speed low."""
-        await self._rpc_attr(
-            self.category_keys["airflow"], self.state_keys["airflow"]["low"]
-        )
-
-    async def medium(self):
-        """Set speed medium."""
-        await self._rpc_attr(
-            self.category_keys["airflow"], self.state_keys["airflow"]["medium"]
-        )
-
-    async def high(self):
-        """Set speed high."""
-        await self._rpc_attr(
-            self.category_keys["airflow"], self.state_keys["airflow"]["high"]
-        )
-
-    async def turbo(self):
-        """Set speed turbo."""
-        await self._rpc_attr(
-            self.category_keys["airflow"], self.state_keys["airflow"]["turbo"]
-        )
-
-    async def sleep(self):
-        """Set device in sleep mode."""
-        await self._rpc_attr(
-            self.category_keys["airflow"], self.state_keys["airflow"]["sleep"]
-        )
-
-    async def _rpc_attr(self, attr: str, value: str):
-        await self._client.get(
-            self.CTRL_URL.format(deviceid=self.device_id, attribute=attr, value=value)
-        )
-        # requests.get(
-        #     self.CTRL_URL.format(deviceid=self.id, attribute=attr, value=value)
-        # )
-
-    async def get_state(self) -> Dict[str, str]:
-        """Get device state."""
-        response = await self._client.get(
-            self.STATE_URL.format(deviceid=self.device_id)
-        )
-        json = await response.json()
-        payload = json["body"]["data"][0]["attributes"]
-        # r = requests.get(self.STATE_URL.format(deviceid=self.id))
-        # payload = r.json()["body"]["data"][0]["attributes"]
-
-        output = dict()
-        for (payload_key, attribute) in payload.items():
-            for (category, local_key) in self.category_keys.items():
-                if payload_key == local_key:
-                    if category in self.state_keys.keys():
-                        for (value_key, value) in self.state_keys[category].items():
-                            if attribute == value:
-                                output[category] = value_key
-                    else:
-                        output[category] = int(attribute)
-
-        return output
