+++ custom_components/winix/__init__.py
@@ -17,7 +17,7 @@ from winix import WinixAccount
 from winix.auth import login, refresh
 from winix.cmd import Configuration
 
-from .WinixDeviceWrapper import WinixDevice, WinixDeviceWrapper
+from .WinixDeviceWrapper import WinixDeviceWrapper
 from .const import DOMAIN, SERVICE_REFRESH_ACCESS
 
 _LOGGER = logging.getLogger(__name__)
