+++ tests/test_driver.py
@@ -1,36 +1,19 @@
 """Test WinixDevice component."""
-from unittest.mock import AsyncMock, Mock, patch
+from unittest.mock import Mock, patch
 
 import pytest
 
-from custom_components.winix import WinixDevice
-from custom_components.winix.const import (
-    AIRFLOW_HIGH,
-    AIRFLOW_LOW,
-    AIRFLOW_SLEEP,
-    ATTR_AIRFLOW,
-    ATTR_MODE,
-    ATTR_PLASMA,
-    ATTR_POWER,
-    MODE_AUTO,
-    MODE_MANUAL,
-    OFF_VALUE,
-    ON_VALUE,
-    PRESET_MODE_AUTO,
-    PRESET_MODE_MANUAL,
-    PRESET_MODE_MANUAL_PLASMA,
-    PRESET_MODE_SLEEP,
-)
+from custom_components.winix.WinixDriver import WinixDriver
 
 
-def build_mock_device() -> WinixDevice:
-    """Returns a mock WinixDeviceWrapper."""
+def build_mock_device() -> WinixDriver:
+    """Return a mocked WinixDeviceWrapper instance."""
     client = Mock()
     device_id = "device_1"
-    return WinixDevice(device_id, client)
+    return WinixDriver(device_id, client)
 
 
-@patch("custom_components.winix.WinixDevice._rpc_attr")
+@patch("custom_components.winix.WinixDriver.WinixDriver._rpc_attr")
 @pytest.mark.parametrize(
     "method, category, value",
     [
@@ -48,12 +31,12 @@ def build_mock_device() -> WinixDevice:
     ],
 )
 async def test_turn_off(rpc_attr, method, category, value):
-    """Test various methods."""
+    """Test various driver methods."""
     device = build_mock_device()
 
     await getattr(device, method)()
     assert rpc_attr.call_count == 1
     assert rpc_attr.call_args[0] == (
-        WinixDevice.category_keys[category],
-        WinixDevice.state_keys[category][value],
+        WinixDriver.category_keys[category],
+        WinixDriver.state_keys[category][value],
     )
